Add files via upload
This Python code uses the Scikit-learn library to build a linear regression model for predicting energy consumption. The code reads in a dataset of energy consumption data from a CSV file, splits the data into training and testing sets, and uses the training set to train a linear regression model. The model is then used to predict the energy consumption for a given date (1/1/2019 in this case) using the test data. The predicted energy consumption value is printed as output.